Add a simple build system
The idea is to make this project simple to include in many web
applications. At the moment I'm adding simple templates for no library
and jquery using applications. I'll probably add prototype at some
point to.

Before this build system is complete I'd like to add some kind of
minification to it and to remove the redundancy between build.sh and
build-jquery.sh