Don't really need the window object anymore
Due to the way the packaging is shaping I think it makes since to just
have one object for the editor and window.

Note: MockCanvas2 in the editorTest is to avoid namespacing issues
with the test canvasDisplayTest. I think its time to start organizing
the test code and start eliminating some of the duplication.