Shift line length to the display
Move the line length into the display object. I believe this number
should be calculated depending on font and canvas size and that
doesn't seem like something that belongs in the editor object/exported
API to me.

At the moment my strategy is to fully test the editor objects API and
work on getting that clean. Everything else should be for supporting
that, so hopefully it will start to organize itself as I go.