Handle unknown keybindings gracefully
When a key is pressed that hasn't been explicitly bound don't throw an
error. We also prevent the default action of the binding. I'm not
completely sure that this is the right behavior, but we don't want
text getting entered into the hidden textarea that has not gone
through the canvas editor interface otherwise they will get out of
sync and you won't be able to tell what is being submitted to the
server.