Add some sensible default bindings.
Just a start to the default keybindings for the editor. Included so
far:

 - a-z and 0-9 with or without Shift held
 - enter
 - backspace
 - space

There are still a bunch of obvious ones to bind, but before doing that
I think its important to write tests to ensure that these are
working. I think these tests need to check the actual character
entered for each key press somehow to make sure that we handle
different browsers that have different keycodes for the same key.