Allow faces to be applied to ranges of characters
A face can be applied to ranges of characters so effectively multiple
faces can be applied to different regions of an editors display. This
is the start of what will enable syntax highlighting etc.

At the moment the default setup will add a red face in the middle of
the text to demonstrate that its working. Obviously this will be
removed soon