Update input to allow raw keycodes
Now keybindings can be set in the inputManager using the raw keycodes
instead of just the characters that they map to in javascript. This
can be used, for example, when a keyboard binds extra keys but
javascript decodes them as extended ascii chars.

The syntax is to escape the character code using '\' but since
javascript automatically escapes backslashes in strings in practice
the declaration will actually be something like 'C-\\195'

Its probably going to be hard to get these key codes to work across
platforms but it still is helpful especially if you are trying to
trigger a specific keyboard. It wouldn't be too tricky to write a
keyboard detection script that asked for user input and dynamically
set the various codes, though this would be shitty from a user perspective