Use default keyboard handling for browser in places
Rather than overriding and trying to interpret every key pressed, let
the default behavior happen if a specific keybinding hasn't been
registered over the top.

This enables a /ton/
  * Weird keyboards
  * Cross browser input (for unregistered keys)
  * Paste without flash! (no copy yet)

This works by attaching a hidden field when the input goes. If the
editor doesn't recognize the keys pressed as a registered binding, the
default behavior is executed upon this hidden field and its results
inserted into the editor. It is then cleared out.